Inline SetupShouldNotIncludeAsyncResult tests
